CitationFileHandling
This code has been used to normalize the citation count. 
Background: As we have seen previously, 90% of the total documents contain 0 to 150 numbers of citations, so there is a chance that enriching the corpus without normalizing it would introduce a bias towards the data having higher citation counts. 
So we have introduced the idea of higher and lower bounding. we adopted the notation “nc-x-y” where ‘nc’ stands for  ‘normalized citations’, “x” denotes the upper bound and “y” signifies the lower bound.
For instance, consider “x=10”. In this scenario, if the total number of words in a citation exceeds 10, we would only consider the top 10 most frequent words present in the citations. 
On the other hand, when “y=10” if the total number of words in a citation falls below 10, we replicate the citation words until the total count reaches 10.