Academic Paper of this project
Abstract—Information retrieval (IR) technology is the foundation of Web-based search engines and is crucial to biomedical research due to its influence on the development of the software that supports literature searches. It is a branch of computer science which deals with how to process documents with free text so they may be quickly retrieved using keywords from a user’s search. The most successful models used for IR are probabilistic. This paper provides an overview of structure, finding, and exploration of information retrieval within the complex information landscape of the TREC COVID benchmark. By utilising the widely used Okapi BM25 information retrieval system, we conducted extensive information searching and analysis. The main objective of our study is to determine if adding citation information from documents to the corpus could improve the baseline score in information retrieval. Our findings showed promising results as including citation details led to noticeable enhancements in the baseline score. This improvement not only made the retrieval process more effective but also emphasised the important role that citation data can play in refining overall retrieval performance.